[Silver III] Title: 정사각형, Time: 44 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1485. 정사각형/정사각형.cc
@@ -1,7 +1,7 @@
 #include <iostream>
 #include <algorithm>
 using namespace std;
-using ll = long long;
+using ll = int;
 
 ll R(pair<ll, ll> a, pair<ll, ll> b) {
 	return (a.first - b.first) * (a.first - b.first) + (a.second - b.second) * (a.second - b.second);
